generalize show landmarks and labels cli tool

+++ show-landmarks.py
@@ -1,8 +1,34 @@
 import cv2
-from video.play.with_hpe import play_with_hpe
+from argparse import ArgumentParser
+from os.path import exists
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
+    parser = ArgumentParser(
+        prog="show-landmarks",
+        description="Play video annotated with HPE landmarks from either MediaPipe or YOLO",
+    )
+    parser.add_argument("path", help="Path to video to play.")
+    parser.add_argument(
+        "--tool",
+        choices=["mp", "yolo"],
+        help="HPE tool to extract landmarks with. Default = mp",
+        default="mp",
+    )
+    args = parser.parse_args()
+
+    video_name = args.path
+    if not exists(video_name):
+        raise Exception(f"Video file {video_name} not found, cannot play.")
+    tool = args.tool
+
+    from src.hpe.mp.play import play_with_hpe as play_with_mp_hpe
+    from src.hpe.yolo.play import play_with_hpe as play_with_yolo_hpe
+
     try:
-        play_with_hpe("./data/samples/NONE/How to Flag - A Climbing Technique for Achieving Balance__500.mp4")
+        if tool == "mp":
+            play_with_mp_hpe(video_name)
+        else:
+            play_with_yolo_hpe(video_name)
     finally:
         cv2.destroyAllWindows()
